+++ data/SettingsData.lua
@@ -1,0 +1,6 @@
+local _, LIB = ...
+
+LIB.INFO_TEXT_HEIGHTS = {
+    compact = 13,
+    default = 27
+}