+++ locale/localization.deDE.lua
@@ -6,7 +6,7 @@ local L = LIB.Localization
 
 -- Dialog
 
-L["dialog.link.text"] = "Um den Link zu kopieren drücke STRG + C."
+L["dialog.link.text"] = "Um den Link zu kopieren, drücke STRG + C."
 
 -- Profiles
 
@@ -18,8 +18,8 @@ L["profiles.switch.name"] = "Profiltyp wechseln"
 L["profiles.switch.tooltip"] = "Wechselt zwischen Accountprofil und Charakterprofil."
 L["profiles.switch.button.account-to-character"] = "Charakterprofil nutzen"
 L["profiles.switch.button.character-to-account"] = "Accountprofil nutzen"
-L["profiles.switch.confirm.account-to-character"] = "Dieser Charakter nutzt danach ein eigenes Charakterprofil. Diese Änderung erfordert einen UI-Reload. Möchtest du fortfahren?"
-L["profiles.switch.confirm.character-to-account"] = "Dieser Charakter nutzt danach das Accountprofil. Diese Änderung erfordert einen UI-Reload. Möchtest du fortfahren?"
+L["profiles.switch.confirm.account-to-character"] = "Dieser Charakter nutzt danach ein eigenes Charakterprofil. Diese Änderung erfordert ein Neuladen der Benutzeroberfläche. Möchtest du fortfahren?"
+L["profiles.switch.confirm.character-to-account"] = "Dieser Charakter nutzt danach das Accountprofil. Diese Änderung erfordert ein Neuladen der Benutzeroberfläche. Möchtest du fortfahren?"
 L["profiles.delete-character-profiles.name"] = "Charakterprofile zurücksetzen"
 L["profiles.delete-character-profiles.button"] = "Charakterprofile löschen"
 L["profiles.delete-character-profiles.tooltip"] = "Löscht alle Charakterprofile und setzt alle Charaktere auf das Accountprofil."