+++ modules/Profiles.lua
@@ -1,4 +1,4 @@
-local addonName, LIB = ...
+local _, LIB = ...
 
 local L = LIB.Localization
 
@@ -20,11 +20,11 @@ end
 
 --- Returns the localized profile mode text.
 ---
---- @param useAccount boolean Whether the account-wide profile mode is active.
+--- @param useAccountProfile boolean Whether the account-wide profile mode is active.
 ---
 --- @return string text The localized profile mode text.
-function ArcaneWizardLibrary.Profiles:GetModeText(useAccount)
-    if useAccount then
+function ArcaneWizardLibrary.Profiles:GetModeText(useAccountProfile)
+    if useAccountProfile then
         return self:GetText("mode.account")
     end
 
@@ -33,11 +33,11 @@ end
 
 --- Returns the localized text for the profile switch button.
 ---
---- @param useAccount boolean Whether the account-wide profile mode is active.
+--- @param useAccountProfile boolean Whether the account-wide profile mode is active.
 ---
 --- @return string text The localized profile switch button text.
-function ArcaneWizardLibrary.Profiles:GetSwitchButtonText(useAccount)
-    if useAccount then
+function ArcaneWizardLibrary.Profiles:GetSwitchButtonText(useAccountProfile)
+    if useAccountProfile then
         return self:GetText("switch.button.account-to-character")
     end
 
@@ -46,11 +46,11 @@ end
 
 --- Returns the localized confirmation text for switching profile mode.
 ---
---- @param useAccount boolean Whether the account-wide profile mode is active.
+--- @param useAccountProfile boolean Whether the account-wide profile mode is active.
 ---
 --- @return string text The localized profile switch confirmation text.
-function ArcaneWizardLibrary.Profiles:GetSwitchConfirmText(useAccount)
-    if useAccount then
+function ArcaneWizardLibrary.Profiles:GetSwitchConfirmText(useAccountProfile)
+    if useAccountProfile then
         return self:GetText("switch.confirm.account-to-character")
     end